+++ models/squencial_recommender/sasrec.py
@@ -128,7 +128,7 @@ class SASRec(nn.Module):
     
     def forward(self, interactions):
         # his_seqs: [batch_size, seq_len], next_items: [batch_size], next_neg_items: [batch_size]
-        his_seqs = interactions["his_seqs"]
+        his_seqs = interactions["his_seqs"].to(torch.long)
         next_items = interactions["next_items"].to(torch.long)
         next_neg_items = interactions.get("next_neg_items", None)
         if next_neg_items is not None:
@@ -153,7 +153,7 @@ class SASRec(nn.Module):
     
     def inference(self, interactions):
         # his_seqs: [batch_size, seq_len]
-        his_seqs = interactions["his_seqs"]
+        his_seqs = interactions["his_seqs"].to(torch.long)
 
         target_indices = (his_seqs == self.pad_idx).sum(dim=-1) - 1
         his_emb = self.encode_seqs(his_seqs)
@@ -163,7 +163,7 @@ class SASRec(nn.Module):
 
     def predict(self, interactions):
         # his_seqs: [batch_size, seq_len], test_items: [batch_size]
-        his_seqs = interactions["his_seqs"]
+        his_seqs = interactions["his_seqs"].to(torch.long)
         test_items = interactions["test_items"].to(torch.long)
         target_indices = (his_seqs == self.pad_idx).sum(dim=-1) - 1
         his_emb = self.encode_seqs(his_seqs)
